requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
httpx==0.25.2
pydantic==2.5.2
pyyaml==6.0.1
cryptography==41.0.7
